Added a js file and added the get functionality
The get functionality is not perfected. The function still needs to pull individual data from the database and then present it in the HTML page. Personally, I am still learning how to do it, but here is some work that I have done. There is also a lot of commented out information in the HTML and js files. This is because I had different ways of doing it, but because we have our HTML page setup the way it is, we cannot do what those youtube videos did, but the code is there for reference.